chore(EMS-2053): move some account e2e tests into directories
