+++ e2e-tests/insurance/cypress/e2e/journeys/account/sign-in/account-creation-and-sign-in-with-eligibility-answers.spec.js
@@ -1,6 +1,6 @@
-import dashboardPage from '../../../../../pages/insurance/dashboard';
-import { ROUTES } from '../../../../../constants';
-import header from '../../../../../partials/header';
+import dashboardPage from '../../../../../../pages/insurance/dashboard';
+import { ROUTES } from '../../../../../../constants';
+import header from '../../../../../../partials/header';
 
 const { table } = dashboardPage;
